\name{vectorfuncts-package}
\alias{vectorfuncts-package}
\alias{vectorfuncts}
\docType{package}
\title{
\packageTitle{vectorfuncts}
}
\description{
\packageDescription{vectorfuncts}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vectorfuncts}
\packageIndices{vectorfuncts}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{vectorfuncts}

Maintainer: \packageMaintainer{vectorfuncts}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
